requirements.txt
# FinFlow Dependencies
anthropic>=0.40.0
yfinance>=0.2.0
matplotlib>=3.9.0
mplfinance>=0.12.0
flask>=3.0.0
python-telegram-bot>=21.0
finnhub-python>=2.4.0
pandas>=2.2.0
numpy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
